+++ lib/bundler/installer.rb
@@ -3,6 +3,10 @@ require 'rubygems/dependency_installer'
 
 module Bundler
   class Installer < Environment
+    class << self
+      attr_accessor :post_install_messages
+    end
+
     def self.install(root, definition, options = {})
       installer = new(root, definition)
       installer.run(options)
@@ -10,6 +14,14 @@ module Bundler
     end
 
     def run(options)
+      # Create the BUNDLE_PATH directory
+      begin
+        Bundler.bundle_path.mkpath unless Bundler.bundle_path.exist?
+      rescue Errno::EEXIST
+        raise PathError, "Could not install to path `#{Bundler.settings[:path]}` " +
+          "because of an invalid symlink. Remove the symlink so the directory can be created."
+      end
+
       if Bundler.settings[:frozen]
         @definition.ensure_equivalent_gemfile_and_lockfile(options[:deployment])
       end
@@ -35,12 +47,10 @@ module Bundler
           @definition.resolve_remotely!
       end
 
-      # Ensure that BUNDLE_PATH exists
-      Bundler.mkdir_p(Bundler.bundle_path) unless File.exist?(Bundler.bundle_path)
-
       # Must install gems in the order that the resolver provides
       # as dependencies might actually affect the installation of
       # the gem.
+      Installer.post_install_messages = {}
       specs.each do |spec|
         spec.source.fetch(spec) if spec.source.respond_to?(:fetch)
 
@@ -49,13 +59,9 @@ module Bundler
         #   next
         # end
 
-        begin
-          old_args = Gem::Command.build_args
-          Gem::Command.build_args = [Bundler.settings["build.#{spec.name}"]]
+        Bundler.rubygems.with_build_args [Bundler.settings["build.#{spec.name}"]] do
           spec.source.install(spec)
           Bundler.ui.debug "from #{spec.loaded_from} "
-        ensure
-          Gem::Command.build_args = old_args
         end
 
         Bundler.ui.info ""